rewriting pca for consistency
