+++ dpbench/benchmarks/pca/__init__.py
@@ -2,6 +2,7 @@
 #
 # SPDX-License-Identifier: Apache 2.0
 
+from .pca_dpnp import pca as pca_dpnp
 from .pca_initialize import initialize
 from .pca_numba_dpex_n import pca as pca_numba_dpex_n
 from .pca_numba_n import pca as pca_numba_n
@@ -11,6 +12,7 @@ from .pca_numpy import pca as pca_numpy
 __all__ = [
     "initialize",
     "pca_numpy",
+    "pca_dpnp",
     "pca_numba_n",
     "pca_numba_np",
     "pca_numba_dpex_n",
