+++ dpbench/benchmarks/pca/pca_dpnp.py
@@ -2,26 +2,12 @@
 #
 # SPDX-License-Identifier: Apache-2.0
 
-import numba as nb
-import numpy as np
+import dpnp as np
 
 
-@nb.njit(parallel=True, fastmath=True)
-def mean_axis_0(data):
-    tdata = data.T
-    m = np.empty(tdata.shape[0])
-    for i in nb.prange(tdata.shape[0]):
-        sum = 0.0
-        for j in range(tdata.shape[1]):
-            sum += tdata[i, j]
-        m[i] = sum / tdata.shape[1]
-    return m
-
-
-@nb.njit(parallel=True, fastmath=True)
 def pca(data, dims_rescaled_data=2):
-    # mean center the data (data -= data.mean(axis=0))
-    data = data - mean_axis_0(data)
+    # mean center the data
+    data -= data.mean(axis=0)
 
     # calculate the covariance matrix
     v = np.cov(data, rowvar=False)
