+++ dpbench/benchmarks/pca/pca_initialize.py
@@ -6,4 +6,4 @@
 def initialize(npoints, dims):
     from sklearn.datasets import make_regression
 
-    return make_regression(n_samples=npoints, n_features=dims)
+    return make_regression(n_samples=npoints, n_features=dims, random_state=0)
